negate mask
