+++ src/experiments/mingru_config.py
@@ -7,7 +7,7 @@ class MinGRUConfig:
     Configuration for minGRU model.
     """
     name: str = 'mingru'
-    embedding_dim: int = 768
+    embedding_dim: int = 128
     expansion_factor: float = 2.5
-    num_layers: int = 4
+    num_layers: int = 2
     bidirectional: bool = True
